Use c++11 strongly typed enums almost everywhere
They have nice misuse-resistant properties

Kept as-is for now: opcua and asn1 serialization

+++ src/aktualizr_secondary/aktualizr_secondary.cc
@@ -139,7 +139,7 @@ bool AktualizrSecondary::sendFirmwareResp(const std::string& firmware) {
         OstreeManager::pull(config_.pacman.sysroot, treehub_server, keys_, target_->sha256Hash());
 
     if (res_code != data::UpdateResultCode::OK) {
-      LOG_ERROR << "Could not pull from OSTree (" << res_code << "): " << message;
+      LOG_ERROR << "Could not pull from OSTree (" << static_cast<int>(res_code) << "): " << message;
       return false;
     }
 #else
@@ -150,7 +150,7 @@ bool AktualizrSecondary::sendFirmwareResp(const std::string& firmware) {
 
   std::tie(res_code, message) = pacman->install(*target_);
   if (res_code != data::UpdateResultCode::OK) {
-    LOG_ERROR << "Could not install target (" << res_code << "): " << message;
+    LOG_ERROR << "Could not install target (" << static_cast<int>(res_code) << "): " << message;
     return false;
   }
